test(editor): keep host command wiring inside the source length budget
Format the host smoke oracle import the way oxfmt does, and move the host
command wiring out of the files the source length gate protects: extension.ts
now registers the gated commands through a small host-test-commands module and
delegates the pure version parser to extension-core, while the host command and
oracle behavior tests live in their own tooling test file.
